Add sanitization tests for inline embeds
